Extract all git operations into GitOperations concern
Consolidate all git-related functionality into a single GitOperations
concern for better organization and maintainability:

- Rename GitCredentialHelper to GitOperations
- Move clone_repository from Run model to concern
- Move push_changes_to_branch from Task model to concern
- Move fetch_branches from Task model to concern
- Move capture_repository_state from Run model to concern
- Add common run_git_command method to reduce duplication
- Make methods work correctly when called from either Run or Task
- Update tests to reflect the new structure

This creates a single place for all git operations, making the code
easier to maintain and extend when adding support for other git
platforms in the future.
