+++ test/models/concerns/git_operations_test.rb
@@ -1,0 +1,77 @@
+require "test_helper"
+
+class GitOperationsTest < ActiveSupport::TestCase
+  setup do
+    Task.any_instance.stubs(:branches).returns([])
+  end
+
+  test "clone_repository uses git credentials for GitHub URLs" do
+    task = tasks(:without_runs)
+    user = task.user
+    user.update!(github_token: "test_token_123")
+
+    project = task.project
+    project.update!(repository_url: "https://github.com/test/repo.git")
+
+    run = task.runs.create!(prompt: "test")
+
+    Docker::Container.expects(:create).with do |config|
+      assert_includes config["Env"], "GITHUB_TOKEN=test_token_123"
+      assert_includes config["Env"], "GIT_ASKPASS=/tmp/git-askpass.sh"
+      assert_match(/git clone https:\/\/github.com\/test\/repo.git/, config["Cmd"][1])
+      true
+    end.returns(mock_container)
+
+    run.clone_repository
+  end
+
+  test "push_changes_to_branch uses git credentials for GitHub URLs" do
+    task = tasks(:without_runs)
+    user = task.user
+    user.update!(github_token: "test_token_123")
+
+    project = task.project
+    project.update!(repository_url: "https://github.com/test/repo.git")
+    task.update!(auto_push_enabled: true, auto_push_branch: "main")
+    task.workplace_mount
+
+    Docker::Container.expects(:create).with do |config|
+      assert_includes config["Env"], "GITHUB_TOKEN=test_token_123"
+      assert_includes config["Env"], "GIT_ASKPASS=/tmp/git-askpass.sh"
+      true
+    end.returns(mock_container)
+
+    task.push_changes_to_branch
+  end
+
+  test "git operations do not use credentials for non-GitHub URLs" do
+    task = tasks(:without_runs)
+    user = task.user
+    user.update!(github_token: "test_token_123")
+
+    project = task.project
+    project.update!(repository_url: "https://gitlab.com/test/repo.git")
+
+    run = task.runs.create!(prompt: "test")
+
+    Docker::Container.expects(:create).with do |config|
+      env = config["Env"] || []
+      refute_includes env, "GITHUB_TOKEN=test_token_123"
+      refute_includes env, "GIT_ASKPASS=/tmp/git-askpass.sh"
+      true
+    end.returns(mock_container)
+
+    run.clone_repository
+  end
+
+  private
+
+  def mock_container
+    container = mock("container")
+    container.expects(:start)
+    container.expects(:wait).returns({ "StatusCode" => 0 })
+    container.expects(:logs).returns("Success")
+    container.expects(:delete)
+    container
+  end
+end
